% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_india_map.R
\name{get_india_map}
\alias{get_india_map}
\title{Get India Map Data or Plot}
\usage{
get_india_map(
  fill,
  location_category,
  location_name = "INDIA",
  action = c("plot_map", "get_map_data")
)
}
\arguments{
\item{fill}{Type of administrative boundary to retrieve ("state", "district", "subdistrict", etc.).}

\item{location_category}{Category of administrative division ("STATE", "DISTRICT", "SUBDISTRICT", etc.).}

\item{location_name}{Name of the administrative division to retrieve (Default: "INDIA").}

\item{action}{Action to perform: "plot_map" (default) or "get_map_data".}
}
\value{
Depending on action, returns spatial data or plots it.
}
\description{
Retrieves map data or plots for Indian administrative boundaries.
}
\details{
This function retrieves Indian administrative boundary data from shapefiles included in the R package "RUtilHub".
}
\examples{
\dontrun{
# Example usage to plot the map for a state
get_india_map(fill = "state", location_category = "STATE", location_name = "karnataka", action = "plot_map")

# Example usage to return the filtered data for a subdistrict
map_data <- get_india_map(fill = "subdistrict", location_category = "STATE", location_name = "UTTARAKHAND", action = "get_map_data")
print(map_data)

# Example usage to plot the map for the entire country
get_india_map(fill = "STATE", location_category = "COUNTRY", location_name = "INDIA", action = "plot_map")
}
}
\seealso{
\code{\link[sf]{st_read}}, \code{\link[sf]{st_union}}
}
