% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_long.R
\name{get_lat_long}
\alias{get_lat_long}
\title{Get Latitude and Longitude for a List of Locations}
\usage{
get_lat_long(locations)
}
\arguments{
\item{locations}{A character vector of locations.}
}
\value{
A data frame with columns `location`, `latitude`, and `longitude`.
}
\description{
This function retrieves the latitude and longitude for a list of locations
using the Nominatim API.
}
\details{
This function queries the Nominatim API for each location provided in the input vector.
It includes error handling to manage failed API requests and respects rate limits by adding a delay
between requests.
}
\examples{
\dontrun{
if(interactive()){
 locations <- c("New York", "Los Angeles", "Chicago")
 lat_long_df <- get_lat_long(locations)
 print(lat_long_df)
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{status_code}}, \code{\link[httr]{content}}
}
